Logical Operators
Implemented and, or, not. These operators do not short-circuit.
And, or occupy a LogicalOpNode, while not is a NotOpNode (being unary).

Added a basic test for them.

For reasons unknown, ANTLR gets confused if and/or/not tokens are valid identifiers, depite derivation ordering. For the moment, the operators are the C# forms & | !.

Dryed out BooleanOpNode by subclassing BinaryOpNode. LogicalOpNode is in turn a subclass of BooleanOpNode.

Now uses the bool CIL type so booleans print True instead of 1.
Related tests changed accordingly.

 Changes to be committed:
	modified:   src/Ast/AstEvaluator.cs
	modified:   src/Ast/AstGenerator.cs
	modified:   src/Ast/AstGeneratorUtils.cs
	modified:   src/Ast/AstNodes.cs
	modified:   src/Ast/TypeResolver.cs
	modified:   src/Blaise.g4
	modified:   src/DotRenderer.cs
	modified:   src/Emitters/BlaiseTypeExtensions.cs
	modified:   src/Emitters/CilEmitter.cs
	modified:   src/Emitters/CilEmitterUtils.cs
	modified:   src/Program.cs
	modified:   tests/CompilerCodeSamplesAttribute.cs
	modified:   tests/CompilerTests.cs