Tinkering and Refactoring
ForLoopNodes no longer have an explicit iterator child node. Instead, AstGenerator combines it into the loop's body node, either appending it to a block or, if it's a single statement, making a new block for the stat and the iteration.
VisitForDo becomes substantially leaner, as both the Iterator build and the block combination move to utility functions.

FunctionReturnEvaluator now uses patern-match dispatch.

Emission of For loops greatly simplified, now just emits the assignment CIL then revisits itself as a LoopNode to reuse that code. It's pleasantly DRY.

Added an emitter utility that checks whether a statement returns and emits a branch to a provided label only if it does not return (along all paths). This prevents branches from being emitted in unreachable positions.

Also streamlines emission of if statements, including an added utility that emits the end label itself only if the ThenStat does not return. This allowed emission of the whole ifstat in one piece rather than the previous, less readable logic.

 Changes to be committed:
	modified:   src/Ast/AstEvaluation/AstEvaluator.cs
	modified:   src/Ast/AstEvaluation/FunctionReturnEvaluator.cs
	modified:   src/Ast/AstGenerator.cs
	modified:   src/Ast/AstNodes/AstNodes.cs
	modified:   src/Emitters/CilEmitter.cs
	modified:   src/Emitters/CilEmitterUtils.cs