Non-dynamic Ast Folding
Found a way to de-dynamic the IConstantNode value returns.
New class: AstConstant, a box that holds either a bool, int, double, char, or string. Has an enum field specifying what kind of value it is. Has a constructor and a Get method per possible type.
Calling a get method that does not match the held value will cast the value to that type.
This is not checked, and can throw exceptions on bad casts.
Use of this class relies on knowledge that the input has passed the AstEvaluator validity checks.

Also defined an IComparer for AstConstants, so I don't have to extract the value at all to do comparison operations on them.

 Changes to be committed:
	new file:   src/Ast/AstConstant.cs
	modified:   src/Ast/AstEvaluation/AstFolder.cs
	modified:   src/Ast/AstNodes/AstNodes.cs
	modified:   src/Ast/AstNodes/IConstantNode.cs